moving tests for another AnotherAssembly to their own file
