+++ Assets/Mirror/Tests/Editor/Weaver/WeaverGeneratedReaderWriterAnotherAssemblyTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a2fc12fbf5fb9274dbeed361445f6cf3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
